+++ src/types.ts
@@ -8,6 +8,7 @@ import BN from 'bn.js';
 import dayjs from 'dayjs';
 import { EAssets } from './components/OpenGovTreasuryProposal/types';
 import { IBountyListing } from './components/Bounties/BountiesListing/types/types';
+import type { RegistrationJudgement } from '@polkadot/types/interfaces';
 
 declare global {
 	interface Window {
@@ -372,7 +373,6 @@ export enum EAllowedCommentor {
 	ONCHAIN_VERIFIED = 'onchain_verified',
 	NONE = 'none'
 }
-
 export interface Post {
 	user_id: number;
 	content: string;
@@ -1096,4 +1096,22 @@ export enum EExpertReqStatus {
 	APPROVED = 'approved',
 	REJECTED = 'rejected',
 	PENDING = 'pending'
+}
+export interface IIdentityInfo {
+	display: string;
+	legal: string;
+	email: string;
+	twitter: string;
+	web: string;
+	github: string;
+	discord: string;
+	matrix: string;
+	displayParent: string;
+	nickname: string;
+	isIdentitySet: boolean;
+	isVerified: boolean;
+	isGood: boolean;
+	judgements: RegistrationJudgement[];
+	verifiedByPolkassembly: boolean;
+	parentProxyTitle: string | null;
 }
