+++ src/ui-components/Address.tsx
@@ -15,7 +15,7 @@ import getEncodedAddress from '~src/util/getEncodedAddress';
 import { getKiltDidName } from '~src/util/kiltDid';
 import shortenAddress from '~src/util/shortenAddress';
 import EthIdenticon from './EthIdenticon';
-import { EAddressOtherTextType } from '~src/types';
+import { EAddressOtherTextType, IIdentityInfo } from '~src/types';
 import classNames from 'classnames';
 import styled from 'styled-components';
 import IdentityBadge from './IdentityBadge';
@@ -31,7 +31,7 @@ import Image from 'next/image';
 import { isAddress } from 'ethers';
 import { dmSans } from 'pages/_app';
 import SkeletonAvatar from '~src/basic-components/Skeleton/SkeletonAvatar';
-import getIdentityInformation, { IIdentityInfo } from '~src/auth/utils/getIdentityInformation';
+import getIdentityInformation from '~src/auth/utils/getIdentityInformation';
 import { usePeopleChainApiContext } from '~src/context';
 import isPeopleChainSupportedNetwork from '~src/components/OnchainIdentity/utils/getPeopleChainSupportedNetwork';
 import copyToClipboard from '~src/util/copyToClipboard';
